added dnacentersdk fork, added templateParms

+++ scripts/dnac_templ.py
@@ -50,32 +50,45 @@ class DNACTemplate(object):
 
     def retrieve_provisioned_templates(self):
         '''
-        retrieve a list of Jinja2 templates currently provisioned.
+        retrieve a list of templates currently provisioned.
         Returns a dict of template detail dicts, indexed by name
         '''
         logger.debug('Retrieving existing templates')
         result = {}
         for t in self.dnac.configuration_templates.gets_the_templates_available(
                 project_id=self.template_project_id):
-            # store both template and template details, joining both in the
-            # same dict
+            # store both template and template details, joining both in the same dict
             result[t.name] = t
             result[t.name].update(self.dnac.configuration_templates.get_template_details(t.templateId))
             logger.debug('Retrieved template {}, full info: {}'.format(t.templateId, result[t.name]))
         return result
 
-    def _retrieve_variables(self, content):
-        # TODO
+    def get_template_params(self, content):
+        '''
+        extracts referenced jinja2 variables and returns params list.
+        TODO: This is not robust, we assume all variables found are strings
+        '''
         from jinja2 import Environment, PackageLoader, meta
         env = Environment(loader=PackageLoader('gummi', 'templates'))
         parsed_content = env.parse(content)
-        meta.find_undeclared_variables(parsed_content)
-        breakpoint()
+        variables = meta.find_undeclared_variables(parsed_content)
+        params = []
+        order = 1
+        for v in variables:
+            params.append({
+                'parameterName': v,
+                'dataType': 'STRING',
+                'required': True,
+                'order': order,
+                'customOrder': 0})
+            order += 1
+            # d ={'parameterName': 'password', 'dataType': 'STRING', 'defaultValue': None,
+            # 'description': None, 'required': True, 'notParam': False, 'paramArray': False,
+            # 'displayName': None, 'instructionText': None, 'group': None, 'order': 2,
+            # 'customOrder': 0, 'selection': None, 'range': [], 'key': None, 'provider':
+            # None, 'binding': '', 'id': '869b0b95-8831-4e9e-9480-5268755bb270'
 
-    def get_template_params(self, content):
-        # TODO
-        # self._retrieve_variables(content)
-        return []
+        return params
 
     def wait_and_check_status(self, response, max_attempts=2, sleeptime=2):
         attempt = 0
@@ -93,7 +106,7 @@ class DNACTemplate(object):
             raise Exception()
         return result
 
-    def provision_templates(self):
+    def provision_templates(self, purge=True):
         '''
         Push all templates found in our git repo to DNAC
         TODL Templates which have previously provisioned but which have been removed
@@ -115,52 +128,53 @@ class DNACTemplate(object):
 
         # process all the templates found in the repo
         for template_file in os.listdir(self.template_dir):
-            logger.debug('processing file ' + template_file)
+            logger.debug('processing file "{}"'.format(template_file))
 
             with open(os.path.join(self.template_dir, template_file), 'r') as fd:
                 template_content = fd.read()
 
-            # self.get_template_params(template_content)
-            # continue
-            #
+            template_name = template_file
 
-            current_template = provisioned_templates.get(template_file)
+            current_template = provisioned_templates.get(template_name)
             if not current_template:
                 # new template
                 params = {
                     'project_id': self.template_project_id,
-                    'name': template_file,
+                    'name': template_name,
+                    'description': '',
+                    'containingTemplates': [],
                     'language': 'JINJA',
                     'composite': False,
                     'deviceTypes': [{'productFamily': 'Switches and Hubs'}],
                     'softwareType': "IOS-XE",
+                    'softwareVersion': None,
+                    'tags': [],
                     'templateParams': self.get_template_params(template_content),
                     'templateContent': template_content
                 }
                 # create the template
-                logger.info('Creating template {}'.format(template_file))
+                logger.info('Creating template "{}"'.format(template_name))
                 logger.debug(params)
                 response = self.dnac.configuration_templates.create_template(**params)
 
             else:
                 # check if content changed
                 if template_content == current_template.templateContent:
-                    logger.info('No change in template {}, no update needed'.format(template_file))
+                    logger.info('No change in template "{}", no update needed'.format(template_name))
                     # mark it so we don't delete it at the end
-                    pushed_templates.append(template_file)
+                    pushed_templates.append(template_name)
                     continue
 
-                logger.info('Updating template {}'.format(template_file))
+                logger.info('Updating template "{}"'.format(template_name))
                 params = {
                     'id': current_template.id,
                     'projectId': self.template_project_id,
-                    'name': template_file,
+                    'name': template_name,
                     'language': current_template.language,
                     'composite': current_template.composite,
                     'softwareType': current_template.softwareType,
                     'deviceTypes': current_template.deviceTypes,
-                    # 'templateParams': self.get_template_params(template_content),
-                    'templateParams': current_template.templateParams,
+                    'templateParams': self.get_template_params(template_content),
                     'templateContent': template_content
                 }
                 logger.debug(params)
@@ -169,7 +183,7 @@ class DNACTemplate(object):
             # check task and retrieve the template_id
             template_id = self.wait_and_check_status(response)
             if not template_id:
-                raise Exception('Creation of template {} failed'.format(template_file))
+                raise Exception('Creation of template "{}" failed'.format(template_name))
 
             # Commit the template
             response = self.dnac.configuration_templates.version_template(
@@ -177,14 +191,17 @@ class DNACTemplate(object):
                 comments='committed by gitlab-ci at {} UTC'.format(datetime.utcnow()))
             self.wait_and_check_status(response)
 
-            pushed_templates.append(template_file)
+            pushed_templates.append(template_name)
 
         # now that we processed all templates, check if there are any
         # templates left on DNAC, which we will delete
         for k, v in provisioned_templates.items():
             if k not in pushed_templates:
-                logger.info('deleting template {}'.format(k))
-                self.dnac.configuration_templates.delete_template(v.id)
+                if purge is True:
+                    logger.info('deleting template "{}"'.format(k))
+                    self.dnac.configuration_templates.delete_template(v.id)
+                else:
+                    logger.info('Not attempting to purge template "{}"'.format(k))
 
         return errors == 0
 
@@ -193,6 +210,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Provision DNAC templates')
     parser.add_argument('--debug', action='store_true', help='print more debugging output')
     parser.add_argument('--config', help='config file to use')
+    parser.add_argument('--nopurge', action="store_true", help='Don\'t delete templates found on DNAC which are not in the repo')
     args = parser.parse_args()
 
     if args.debug:
@@ -201,5 +219,5 @@ if __name__ == '__main__':
         logging.basicConfig(level=logging.INFO)
 
     service = DNACTemplate(config_file=args.config)
-    result = service.provision_templates()
+    result = service.provision_templates(purge=not args.nopurge)
     sys.exit(0 if result else 1)
